Drafted the undisputed webservices (#86)
